refactor(frontend): split ExperienceList helpers into folder
Extract sub-components from the monolithic ExperienceList.tsx (1344 → 437 lines):
- ExperienceList/ArtworksList.tsx: artwork grid for treasure-bearing experiences
- ExperienceList/ExperienceListItem.tsx: single experience row with expansion
- ExperienceList/ExperienceExpandedDetails.tsx: expanded detail panel
- ExperienceList/VisitedStatusButton.tsx: visit toggle button
- ExperienceList/utils.ts: sorting and grouping utilities

ExperienceList.tsx now imports from the subfolder and is the thin orchestrator.
