+++ frontend/src/components/ExperienceList.tsx
@@ -17,65 +17,36 @@ import {
   ListItem,
   ListItemText,
   ListItemIcon,
-  Checkbox,
   Collapse,
   IconButton,
   Button,
-  Chip,
   Tooltip,
-  Alert,
 } from '@mui/material';
 import {
   ExpandLess,
   ExpandMore,
-  Place as PlaceIcon,
-  CheckCircle as CheckCircleIcon,
-  MenuBook as WikiIcon,
-  Language as WebsiteIcon,
-  IndeterminateCheckBox as PartialIcon,
-  LocationOn as LocationIcon,
   Add as AddIcon,
-  Remove as RemoveIcon,
   Block as RejectIcon,
-  Undo as UnrejectIcon,
   PlaylistAdd as AssignIcon,
-  Tune as CurateIcon,
-  LinkOff as RemoveFromRegionIcon,
 } from '@mui/icons-material';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { useExperienceContext, extractImageUrl, toThumbnailUrl } from '../hooks/useExperienceContext';
+import { useExperienceContext } from '../hooks/useExperienceContext';
 import { useAuth } from '../hooks/useAuth';
-import { useVisitedExperiences, useVisitedLocations, useViewedTreasures } from '../hooks/useVisitedExperiences';
+import { useVisitedExperiences, useVisitedLocations } from '../hooks/useVisitedExperiences';
 import { useRegionLocations } from '../hooks/useRegionLocations';
 import {
-  fetchExperience,
-  fetchExperienceTreasures,
   fetchExperienceCategories,
   unrejectExperience,
   removeExperienceFromRegion,
   type Experience,
-  type ExperienceLocation,
-  type VisitedStatus,
 } from '../api/experiences';
 import { useNavigation } from '../hooks/useNavigation';
 import { CurationDialog } from './shared/CurationDialog';
 import { AddExperienceDialog } from './shared/AddExperienceDialog';
-
-const NEW_BADGE_DAYS = 7;
-const OUT_OF_REGION_INITIAL = 3;
-
-function isNewExperience(createdAt?: string): boolean {
-  if (!createdAt) return false;
-  const created = new Date(createdAt);
-  const cutoff = new Date();
-  cutoff.setDate(cutoff.getDate() - NEW_BADGE_DAYS);
-  return created > cutoff;
-}
-
-import { getCategoryPrimaryColor, VISITED_GREEN, PARTIAL_AMBER } from '../utils/categoryColors';
 import { scrollToCenter, scrollToTop } from '../utils/scrollUtils';
 import { invalidateExperiences } from '../utils/queryInvalidation';
 import { LoadingSpinner } from './shared/LoadingSpinner';
+import { ExperienceListItem } from './ExperienceList/ExperienceListItem';
 
 interface ExperienceGroup {
   categoryName: string;
@@ -462,883 +433,5 @@ export function ExperienceList({ scrollContainerRef }: ExperienceListProps) {
         />
       )}
     </Box>
-  );
-}
-
-// =============================================================================
-// Individual List Item with Expandable Details
-// =============================================================================
-
-interface ExperienceListItemProps {
-  experience: Experience;
-  locations?: ExperienceLocation[];
-  isLocationVisited: (locationId: number) => boolean;
-  isHovered: boolean;
-  isSelected: boolean;
-  hoveredLocationId: number | null;
-  locationRefs: React.MutableRefObject<Map<number, HTMLDivElement>>;
-  showCheckbox: boolean;
-  isLoading: boolean;
-  onHover: () => void;
-  onLeave: () => void;
-  onClick: () => void;
-  onLocationVisitedToggle: (locationId: number, isVisited: boolean) => void;
-  onToggleAllLocations: (experienceId: number, markAsVisited: boolean) => void;
-  onLocationHover: (locationId: number | null) => void;
-  onCurate?: () => void;
-  onUnreject?: () => void;
-  onRemoveFromRegion?: () => void;
-  isRejected?: boolean;
-}
-
-function ExperienceListItem({
-  experience,
-  locations,
-  isLocationVisited,
-  isHovered,
-  isSelected,
-  hoveredLocationId,
-  locationRefs,
-  showCheckbox,
-  isLoading,
-  onHover,
-  onLeave,
-  onClick,
-  onLocationVisitedToggle,
-  onToggleAllLocations,
-  onLocationHover,
-  onCurate,
-  onUnreject,
-  onRemoveFromRegion,
-  isRejected,
-}: ExperienceListItemProps) {
-  const color = getCategoryPrimaryColor(experience.category);
-
-  // Use batch locations from parent (shared hook) — no per-item fetch
-  const totalLocations = locations?.length ?? (experience.location_count ?? 0);
-  const inRegionLocations = useMemo(() => {
-    if (!locations) return [];
-    return locations.filter(l => l.in_region !== false);
-  }, [locations]);
-  const inRegionCount = inRegionLocations.length;
-  const isMultiLocation = totalLocations > 1;
-
-  // Compute IN-REGION visited status using global isLocationVisited
-  const inRegionVisitedStatus = useMemo((): 'not_visited' | 'partial' | 'visited' => {
-    if (inRegionCount === 0) return 'not_visited';
-
-    const inRegionVisitedCount = inRegionLocations.filter(loc => isLocationVisited(loc.id)).length;
-
-    if (inRegionVisitedCount === 0) return 'not_visited';
-    if (inRegionVisitedCount >= inRegionCount) return 'visited';
-    return 'partial';
-  }, [inRegionLocations, inRegionCount, isLocationVisited]);
-
-  // Derive checkbox state from in-region locations
-  const isPartiallyVisited = inRegionVisitedStatus === 'partial';
-  const isFullyVisited = inRegionVisitedStatus === 'visited';
-
-  // Handle root checkbox click - always use batch operation (works for single and multi-location)
-  const handleRootCheckboxClick = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    // Toggle all in-region locations: if not all visited, mark all; else unmark all
-    onToggleAllLocations(experience.id, inRegionVisitedStatus !== 'visited');
-  };
-
-  // Batch mark all in-region locations
-  const handleMarkAll = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    onToggleAllLocations(experience.id, true);
-  };
-
-  // Batch unmark all in-region locations
-  const handleUnmarkAll = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    onToggleAllLocations(experience.id, false);
-  };
-
-  return (
-    <Box>
-      {/* Main list item */}
-      <ListItem
-        sx={{
-          pl: 2,
-          cursor: 'pointer',
-          bgcolor: isHovered ? 'action.hover' : isSelected ? 'primary.50' : 'transparent',
-          borderLeft: isSelected ? 4 : 0,
-          borderColor: 'primary.main',
-          '&:hover': { bgcolor: 'action.hover' },
-          borderBottom: isSelected ? 0 : '1px solid',
-          borderBottomColor: 'divider',
-        }}
-        onMouseEnter={onHover}
-        onMouseLeave={onLeave}
-        onClick={onClick}
-      >
-        {/* Checkbox for visited status + batch buttons when partial */}
-        {showCheckbox && (
-          <ListItemIcon sx={{ minWidth: isPartiallyVisited ? 72 : 36, display: 'flex', alignItems: 'center' }}>
-            <Checkbox
-              edge="start"
-              checked={isFullyVisited}
-              indeterminate={isPartiallyVisited}
-              disabled={isLoading}
-              onClick={handleRootCheckboxClick}
-              sx={{
-                color: color,
-                '&.Mui-checked': { color: VISITED_GREEN },
-                '&.MuiCheckbox-indeterminate': { color: PARTIAL_AMBER }, // Amber for partial
-              }}
-            />
-            {/* Batch buttons when partial - mark all / unmark all */}
-            {isPartiallyVisited && (
-              <Box sx={{ display: 'flex', ml: -0.5 }}>
-                <Tooltip title="Mark all in region">
-                  <IconButton
-                    size="small"
-                    onClick={handleMarkAll}
-                    disabled={isLoading}
-                    sx={{ p: 0.25, color: '#10B981' }}
-                  >
-                    <AddIcon sx={{ fontSize: 16 }} />
-                  </IconButton>
-                </Tooltip>
-                <Tooltip title="Unmark all in region">
-                  <IconButton
-                    size="small"
-                    onClick={handleUnmarkAll}
-                    disabled={isLoading}
-                    sx={{ p: 0.25, color: '#EF4444' }}
-                  >
-                    <RemoveIcon sx={{ fontSize: 16 }} />
-                  </IconButton>
-                </Tooltip>
-              </Box>
-            )}
-          </ListItemIcon>
-        )}
-
-        {/* Category color indicator */}
-        <ListItemIcon sx={{ minWidth: 32 }}>
-          <PlaceIcon sx={{ color, fontSize: 20 }} />
-        </ListItemIcon>
-
-        {/* Experience name */}
-        <ListItemText
-          primary={
-            <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-              <Typography
-                variant="body2"
-                sx={{
-                  fontWeight: isSelected ? 600 : 500,
-                  textDecoration: isFullyVisited ? 'line-through' : 'none',
-                  color: isFullyVisited ? 'text.secondary' : isPartiallyVisited ? 'text.primary' : 'text.primary',
-                  flex: 1,
-                }}
-              >
-                {experience.name}
-              </Typography>
-              {isNewExperience(experience.created_at) && (
-                <Chip
-                  label="New"
-                  size="small"
-                  color="success"
-                  sx={{
-                    height: 18,
-                    fontSize: '0.6rem',
-                    fontWeight: 700,
-                    '& .MuiChip-label': { px: 0.5 },
-                  }}
-                />
-              )}
-              {isMultiLocation && (
-                <Chip
-                  label={inRegionCount === totalLocations ? totalLocations : `${inRegionCount}/${totalLocations}`}
-                  size="small"
-                  sx={{
-                    height: 18,
-                    fontSize: '0.65rem',
-                    '& .MuiChip-label': { px: 0.75 },
-                  }}
-                  variant="outlined"
-                  color="info"
-                />
-              )}
-            </Box>
-          }
-        />
-
-        {/* Expand indicator */}
-        {isSelected ? <ExpandLess fontSize="small" /> : <ExpandMore fontSize="small" />}
-      </ListItem>
-
-      {/* Expanded details */}
-      <Collapse in={isSelected} timeout="auto" unmountOnExit>
-        <ExperienceExpandedDetails
-          experience={experience}
-          locations={locations}
-          isLocationVisited={isLocationVisited}
-          isFullyVisited={isFullyVisited}
-          hoveredLocationId={hoveredLocationId}
-          locationRefs={locationRefs}
-          showCheckbox={showCheckbox}
-          onToggleAllLocations={onToggleAllLocations}
-          onLocationVisitedToggle={onLocationVisitedToggle}
-          onLocationHover={onLocationHover}
-          onCurate={onCurate}
-          onUnreject={onUnreject}
-          onRemoveFromRegion={onRemoveFromRegion}
-          isRejected={isRejected}
-        />
-      </Collapse>
-    </Box>
-  );
-}
-
-// =============================================================================
-// Expanded Details Panel
-// =============================================================================
-
-interface ExperienceExpandedDetailsProps {
-  experience: Experience;
-  locations?: ExperienceLocation[];
-  isLocationVisited: (locationId: number) => boolean;
-  isFullyVisited: boolean;
-  hoveredLocationId: number | null;
-  locationRefs: React.MutableRefObject<Map<number, HTMLDivElement>>;
-  showCheckbox: boolean;
-  onToggleAllLocations: (experienceId: number, markAsVisited: boolean) => void;
-  onLocationVisitedToggle: (locationId: number, isVisited: boolean) => void;
-  onLocationHover: (locationId: number | null) => void;
-  onCurate?: () => void;
-  onUnreject?: () => void;
-  onRemoveFromRegion?: () => void;
-  isRejected?: boolean;
-}
-
-function ExperienceExpandedDetails({
-  experience,
-  locations,
-  isLocationVisited,
-  isFullyVisited,
-  hoveredLocationId,
-  locationRefs,
-  showCheckbox,
-  onToggleAllLocations,
-  onLocationVisitedToggle,
-  onLocationHover,
-  onCurate,
-  onUnreject,
-  onRemoveFromRegion,
-  isRejected,
-}: ExperienceExpandedDetailsProps) {
-  const { isAuthenticated } = useAuth();
-
-  // Fetch full details
-  const { data: details } = useQuery({
-    queryKey: ['experience', experience.id],
-    queryFn: () => fetchExperience(experience.id),
-    staleTime: 300000,
-  });
-
-  // Fetch contents (artworks) - only if experience has contents
-  const { data: contentsData } = useQuery({
-    queryKey: ['experience-contents', experience.id],
-    queryFn: () => fetchExperienceTreasures(experience.id),
-    staleTime: 300000,
-  });
-
-  // Use batch locations from parent + global isLocationVisited
-  const totalLocations = locations?.length ?? (experience.location_count ?? 0);
-
-  // Build location display data with visited + in_region info from shared data
-  const locationsWithRegionInfo = useMemo(() => {
-    if (!locations || locations.length === 0) return [];
-    return locations.map(loc => ({
-      id: loc.id,
-      name: loc.name,
-      ordinal: loc.ordinal,
-      longitude: loc.longitude,
-      latitude: loc.latitude,
-      isVisited: isLocationVisited(loc.id),
-      inRegion: loc.in_region !== false,
-      regionPath: loc.region_path ?? null,
-    }));
-  }, [locations, isLocationVisited]);
-
-  // Split into in-region and out-of-region
-  const inRegionLocs = useMemo(
-    () => locationsWithRegionInfo.filter(l => l.inRegion),
-    [locationsWithRegionInfo],
-  );
-  const outOfRegionLocs = useMemo(
-    () => locationsWithRegionInfo.filter(l => !l.inRegion),
-    [locationsWithRegionInfo],
-  );
-  const [outOfRegionExpanded, setOutOfRegionExpanded] = useState(false);
-
-  // Compute display paths: strip common prefix segments shared by all out-of-region locations
-  const outOfRegionDisplayPaths = useMemo(() => {
-    const paths = outOfRegionLocs.map(l => l.regionPath);
-    if (paths.length <= 1) {
-      // Single location or none — show full path
-      return new Map(outOfRegionLocs.map(l => [l.id, l.regionPath]));
-    }
-    // Split into segments and find common prefix length
-    const segmented = paths.map(p => p?.split(' > ') ?? []);
-    const firstSegs = segmented[0];
-    let commonLen = 0;
-    for (let i = 0; i < firstSegs.length; i++) {
-      if (segmented.every(s => s[i] === firstSegs[i])) {
-        commonLen = i + 1;
-      } else {
-        break;
-      }
-    }
-    return new Map(outOfRegionLocs.map((l, idx) => {
-      const segs = segmented[idx];
-      const trimmed = segs.slice(commonLen).join(' > ');
-      return [l.id, trimmed || l.regionPath];
-    }));
-  }, [outOfRegionLocs]);
-
-  // Count in-region locations
-  const inRegionCount = inRegionLocs.length;
-  const inRegionVisitedCount = inRegionLocs.filter(l => l.isVisited).length;
-
-  // Compute visited status for multi-location badge
-  const visitedLocations = locationsWithRegionInfo.filter(l => l.isVisited).length;
-  const visitedStatus: VisitedStatus = visitedLocations === 0
-    ? 'not_visited'
-    : visitedLocations >= totalLocations
-      ? 'visited'
-      : 'partial';
-
-  const imageUrl = extractImageUrl(experience.image_url);
-  const isMultiLocation = totalLocations > 1;
-
-  const categoryColorMap: Record<string, { bg: string; text: string }> = {
-    cultural: { bg: '#EDE9FE', text: '#7C3AED' },
-    natural: { bg: '#D1FAE5', text: '#059669' },
-    mixed: { bg: '#FEF3C7', text: '#D97706' },
-  };
-
-  const categoryStyle = categoryColorMap[experience.category || ''] || { bg: '#E0E7FF', text: '#4F46E5' };
-
-  return (
-    <Box
-      sx={{
-        pl: 2,
-        pr: 2,
-        py: 1.5,
-        bgcolor: 'grey.50',
-        borderBottom: '1px solid',
-        borderColor: 'divider',
-      }}
-    >
-      {/* Image */}
-      {imageUrl && (
-        <Box
-          component="img"
-          src={toThumbnailUrl(imageUrl, 330)}
-          alt={experience.name}
-          sx={{
-            width: '100%',
-            maxHeight: 250,
-            objectFit: 'contain',
-            borderRadius: 1,
-            mb: 2,
-            bgcolor: 'grey.100',
-          }}
-          onError={(e) => {
-            (e.target as HTMLImageElement).style.display = 'none';
-          }}
-        />
-      )}
-
-      {/* Category & Country chips */}
-      <Box sx={{ display: 'flex', gap: 1, mb: 1.5, flexWrap: 'wrap' }}>
-        {experience.category && (
-          <Chip
-            label={experience.category}
-            size="small"
-            sx={{
-              bgcolor: categoryStyle.bg,
-              color: categoryStyle.text,
-              fontWeight: 500,
-              textTransform: 'capitalize',
-            }}
-          />
-        )}
-        {experience.country_names?.[0] && (
-          <Chip label={experience.country_names[0]} size="small" variant="outlined" />
-        )}
-        {experience.in_danger && (
-          <Chip label="In Danger" size="small" color="error" />
-        )}
-        {isMultiLocation && (
-          <Chip
-            label={`${inRegionCount}/${totalLocations} in region`}
-            size="small"
-            icon={<LocationIcon fontSize="small" />}
-            variant="outlined"
-            color="info"
-          />
-        )}
-      </Box>
-
-      {/* Description */}
-      {experience.short_description && (
-        <Typography variant="body2" color="text.secondary" sx={{ mb: 1.5 }}>
-          {experience.short_description}
-        </Typography>
-      )}
-
-      {/* Date inscribed */}
-      {details?.metadata?.dateInscribed != null && (
-        <Typography variant="caption" color="text.secondary" sx={{ display: 'block', mb: 1.5 }}>
-          Inscribed: {String(details.metadata.dateInscribed as string | number)}
-        </Typography>
-      )}
-
-      {/* Museum description */}
-      {details?.description && !experience.short_description && (
-        <Typography variant="body2" color="text.secondary" sx={{ mb: 1.5 }}>
-          {details.description}
-        </Typography>
-      )}
-
-      {/* Artworks / Contents list */}
-      {contentsData && contentsData.treasures.length > 0 && (
-        <ArtworksList contents={contentsData.treasures} total={contentsData.total} experienceId={experience.id} />
-      )}
-
-      {/* Multi-location list */}
-      {isMultiLocation && locationsWithRegionInfo.length > 0 && (
-        <Box sx={{ mb: 2 }}>
-          {isAuthenticated && showCheckbox && (
-            <Typography variant="caption" color="text.secondary" sx={{ mb: 1, display: 'block' }}>
-              In this region: {inRegionVisitedCount}/{inRegionCount} visited
-            </Typography>
-          )}
-          {!isAuthenticated && (
-            <Typography variant="caption" color="text.secondary" sx={{ mb: 1, display: 'block' }}>
-              {inRegionCount} location{inRegionCount !== 1 ? 's' : ''} in this region
-            </Typography>
-          )}
-          <List
-            dense
-            disablePadding
-            sx={{ bgcolor: 'white', borderRadius: 1, border: '1px solid', borderColor: 'divider' }}
-            onMouseLeave={() => onLocationHover(null)}
-          >
-            {/* In-region locations (always shown) */}
-            {inRegionLocs.map((loc) => {
-              const isLocationHovered = hoveredLocationId === loc.id;
-              return (
-                <Box
-                  key={loc.id}
-                  ref={(el: HTMLDivElement | null) => {
-                    if (el) {
-                      locationRefs.current.set(loc.id, el);
-                    } else {
-                      locationRefs.current.delete(loc.id);
-                    }
-                  }}
-                >
-                  <ListItem
-                    dense
-                    sx={{
-                      py: 0.5,
-                      bgcolor: isLocationHovered ? 'primary.100' : 'transparent',
-                      cursor: 'pointer',
-                      '&:hover': { bgcolor: 'action.hover' },
-                      transition: 'background-color 0.15s ease',
-                    }}
-                    onMouseEnter={() => onLocationHover(loc.id)}
-                    secondaryAction={
-                      isAuthenticated && showCheckbox ? (
-                        <Checkbox
-                          edge="end"
-                          checked={loc.isVisited}
-                          size="small"
-                          onChange={() => onLocationVisitedToggle(loc.id, loc.isVisited)}
-                          sx={{
-                            '&.Mui-checked': { color: VISITED_GREEN },
-                          }}
-                        />
-                      ) : undefined
-                    }
-                  >
-                    <ListItemIcon sx={{ minWidth: 28 }}>
-                      <LocationIcon
-                        fontSize="small"
-                        color={isLocationHovered ? 'primary' : 'action'}
-                      />
-                    </ListItemIcon>
-                    <ListItemText
-                      primary={loc.name || `Location ${loc.ordinal + 1}`}
-                      primaryTypographyProps={{
-                        variant: 'body2',
-                        sx: {
-                          textDecoration: loc.isVisited ? 'line-through' : 'none',
-                          color: isLocationHovered
-                            ? 'primary.main'
-                            : loc.isVisited ? 'text.secondary' : 'text.primary',
-                          fontWeight: isLocationHovered ? 600 : 400,
-                        },
-                      }}
-                    />
-                  </ListItem>
-                </Box>
-              );
-            })}
-
-            {/* Out-of-region locations (collapsible) */}
-            {outOfRegionLocs.length > 0 && (
-              <>
-                <Box sx={{ px: 1.5, py: 0.5, bgcolor: 'grey.100', borderTop: '1px solid', borderColor: 'divider' }}>
-                  <Typography variant="caption" color="text.secondary" sx={{ fontWeight: 600 }}>
-                    {outOfRegionLocs.length} outside region
-                  </Typography>
-                </Box>
-                {(outOfRegionExpanded ? outOfRegionLocs : outOfRegionLocs.slice(0, OUT_OF_REGION_INITIAL)).map((loc) => (
-                  <Box
-                    key={loc.id}
-                    ref={(el: HTMLDivElement | null) => {
-                      if (el) {
-                        locationRefs.current.set(loc.id, el);
-                      } else {
-                        locationRefs.current.delete(loc.id);
-                      }
-                    }}
-                  >
-                    <ListItem
-                      dense
-                      sx={{
-                        py: 0.5,
-                        opacity: 0.4,
-                        bgcolor: 'grey.100',
-                        cursor: 'default',
-                      }}
-                    >
-                      <ListItemIcon sx={{ minWidth: 28 }}>
-                        <LocationIcon fontSize="small" color="disabled" />
-                      </ListItemIcon>
-                      <ListItemText
-                        primary={loc.name || `Location ${loc.ordinal + 1}`}
-                        secondary={outOfRegionDisplayPaths.get(loc.id) || 'Outside region'}
-                        primaryTypographyProps={{
-                          variant: 'body2',
-                          sx: { color: 'text.disabled' },
-                        }}
-                        secondaryTypographyProps={{
-                          variant: 'caption',
-                          sx: { fontSize: '0.65rem' },
-                        }}
-                      />
-                    </ListItem>
-                  </Box>
-                ))}
-                {outOfRegionLocs.length > OUT_OF_REGION_INITIAL && (
-                  <Box
-                    sx={{
-                      textAlign: 'center',
-                      py: 0.5,
-                      bgcolor: 'grey.100',
-                      cursor: 'pointer',
-                      '&:hover': { bgcolor: 'grey.200' },
-                    }}
-                    onClick={() => setOutOfRegionExpanded(!outOfRegionExpanded)}
-                  >
-                    <Typography variant="caption" color="primary">
-                      {outOfRegionExpanded
-                        ? 'Show less'
-                        : `Show ${outOfRegionLocs.length - OUT_OF_REGION_INITIAL} more`}
-                    </Typography>
-                  </Box>
-                )}
-              </>
-            )}
-          </List>
-        </Box>
-      )}
-
-      {/* Rejection reason (when viewing rejected item) */}
-      {isRejected && experience.rejection_reason && (
-        <Alert severity="warning" sx={{ mb: 1.5, py: 0 }} variant="outlined">
-          <Typography variant="caption">
-            Rejected: {experience.rejection_reason}
-          </Typography>
-        </Alert>
-      )}
-
-      {/* Actions */}
-      <Box sx={{ display: 'flex', gap: 1, alignItems: 'center', flexWrap: 'wrap' }}>
-        {isAuthenticated && showCheckbox && !isMultiLocation && (
-          <Button
-            variant={isFullyVisited ? 'outlined' : 'contained'}
-            size="small"
-            onClick={(e) => {
-              e.stopPropagation();
-              onToggleAllLocations(experience.id, !isFullyVisited);
-            }}
-            startIcon={isFullyVisited ? <CheckCircleIcon /> : null}
-            color={isFullyVisited ? 'success' : 'primary'}
-          >
-            {isFullyVisited ? 'Visited' : 'Mark Visited'}
-          </Button>
-        )}
-        {isAuthenticated && showCheckbox && isMultiLocation && (
-          <VisitedStatusButton
-            visitedStatus={visitedStatus}
-            visitedCount={visitedLocations}
-            totalCount={totalLocations}
-          />
-        )}
-        {(() => {
-          const metadata = details?.metadata;
-          const wikiUrl = typeof metadata?.wikipediaUrl === 'string' && metadata.wikipediaUrl ? metadata.wikipediaUrl : null;
-          const websiteUrl = typeof metadata?.website === 'string' && metadata.website ? metadata.website : null;
-
-          return (
-            <>
-              {wikiUrl && (
-                <IconButton size="small" component="a" href={wikiUrl} target="_blank" rel="noopener noreferrer" title="Wikipedia">
-                  <WikiIcon fontSize="small" />
-                </IconButton>
-              )}
-              {websiteUrl && websiteUrl !== wikiUrl && (
-                <IconButton size="small" component="a" href={websiteUrl} target="_blank" rel="noopener noreferrer" title="Official website">
-                  <WebsiteIcon fontSize="small" />
-                </IconButton>
-              )}
-            </>
-          );
-        })()}
-
-        {/* Curator actions */}
-        {onCurate && !isRejected && (
-          <Tooltip title="Edit, reject, or manage this experience">
-            <Button
-              size="small"
-              variant="outlined"
-              startIcon={<CurateIcon />}
-              onClick={(e) => { e.stopPropagation(); onCurate(); }}
-              sx={{ ml: 'auto' }}
-            >
-              Curate
-            </Button>
-          </Tooltip>
-        )}
-        {onUnreject && isRejected && (
-          <Button
-            size="small"
-            variant="outlined"
-            color="success"
-            startIcon={<UnrejectIcon />}
-            onClick={(e) => { e.stopPropagation(); onUnreject(); }}
-            sx={{ ml: 'auto' }}
-          >
-            Unreject
-          </Button>
-        )}
-        {onRemoveFromRegion && isRejected && (
-          <Tooltip title="Remove from this region entirely">
-            <IconButton
-              size="small"
-              color="error"
-              onClick={(e) => { e.stopPropagation(); onRemoveFromRegion(); }}
-            >
-              <RemoveFromRegionIcon fontSize="small" />
-            </IconButton>
-          </Tooltip>
-        )}
-        {onCurate && isRejected && (
-          <Tooltip title="Edit or manage this experience">
-            <IconButton
-              size="small"
-              onClick={(e) => { e.stopPropagation(); onCurate(); }}
-            >
-              <CurateIcon fontSize="small" />
-            </IconButton>
-          </Tooltip>
-        )}
-      </Box>
-    </Box>
-  );
-}
-
-// =============================================================================
-// Artworks List (for museum experiences)
-// =============================================================================
-
-const ARTWORKS_INITIAL_LIMIT = 10;
-
-function ArtworksList({ contents, total, experienceId }: { contents: import('../api/experiences').ExperienceTreasure[]; total: number; experienceId: number }) {
-  const { setPreviewImageUrl } = useExperienceContext();
-  const { isAuthenticated } = useAuth();
-  const { viewedIds, viewedCount, markViewed, unmarkViewed } = useViewedTreasures(experienceId);
-  const [showAll, setShowAll] = useState(false);
-  const displayContents = showAll ? contents : contents.slice(0, ARTWORKS_INITIAL_LIMIT);
-  const hasMore = total > ARTWORKS_INITIAL_LIMIT;
-
-  const handleToggleViewed = (treasureId: number, e: React.MouseEvent) => {
-    e.stopPropagation();
-    if (viewedIds.has(treasureId)) {
-      unmarkViewed(treasureId);
-    } else {
-      markViewed({ treasureId, experienceId });
-    }
-  };
-
-  return (
-    <Box sx={{ mb: 2 }}>
-      <Typography variant="caption" color="text.secondary" sx={{ mb: 1, display: 'block', fontWeight: 600 }}>
-        Notable works ({total}){isAuthenticated && viewedCount > 0 && ` \u00B7 ${viewedCount} seen`}
-      </Typography>
-      <Box
-        sx={{
-          bgcolor: 'white',
-          borderRadius: 1,
-          border: '1px solid',
-          borderColor: 'divider',
-          maxHeight: 300,
-          overflowY: 'auto',
-        }}
-      >
-        {displayContents.map((content) => {
-          const isViewed = viewedIds.has(content.id);
-          return (
-            <Box
-              key={content.id}
-              sx={{
-                display: 'flex',
-                alignItems: 'center',
-                gap: 1.5,
-                px: 1.5,
-                py: 1,
-                borderBottom: '1px solid',
-                borderColor: 'divider',
-                '&:last-child': { borderBottom: 0 },
-              }}
-            >
-              {isAuthenticated && (
-                <Checkbox
-                  size="small"
-                  checked={isViewed}
-                  onClick={(e) => handleToggleViewed(content.id, e)}
-                  sx={{
-                    p: 0.25,
-                    flexShrink: 0,
-                    '&.Mui-checked': { color: VISITED_GREEN },
-                  }}
-                />
-              )}
-              {content.image_url && (
-                <Box
-                  sx={{
-                    width: 48,
-                    height: 48,
-                    flexShrink: 0,
-                    display: 'flex',
-                    alignItems: 'center',
-                    justifyContent: 'center',
-                    bgcolor: 'grey.100',
-                    borderRadius: 0.5,
-                    cursor: 'pointer',
-                    opacity: isViewed ? 0.5 : 1,
-                  }}
-                  onMouseEnter={() => setPreviewImageUrl(toThumbnailUrl(content.image_url!, 500))}
-                  onMouseLeave={() => setPreviewImageUrl(null)}
-                >
-                  <Box
-                    component="img"
-                    src={toThumbnailUrl(content.image_url)}
-                    alt={content.name}
-                    loading="lazy"
-                    sx={{
-                      maxWidth: 48,
-                      maxHeight: 48,
-                      objectFit: 'contain',
-                      borderRadius: 0.5,
-                    }}
-                    onError={(e) => {
-                      (e.target as HTMLImageElement).style.display = 'none';
-                    }}
-                  />
-                </Box>
-              )}
-              <Box sx={{ minWidth: 0, flex: 1 }}>
-                <Typography
-                  variant="body2"
-                  sx={{
-                    fontWeight: 500,
-                    lineHeight: 1.3,
-                    textDecoration: isViewed ? 'line-through' : 'none',
-                    color: isViewed ? 'text.secondary' : 'text.primary',
-                  }}
-                  noWrap
-                >
-                  {content.name}
-                </Typography>
-                <Typography variant="caption" color="text.secondary" noWrap>
-                  {[
-                    content.artist,
-                    content.year,
-                    content.treasure_type,
-                  ].filter(Boolean).join(' \u00B7 ')}
-                </Typography>
-              </Box>
-            </Box>
-          );
-        })}
-        {hasMore && !showAll && (
-          <Box
-            sx={{ textAlign: 'center', py: 0.5, cursor: 'pointer', '&:hover': { bgcolor: 'action.hover' } }}
-            onClick={() => setShowAll(true)}
-          >
-            <Typography variant="caption" color="primary">
-              Show all {total} works
-            </Typography>
-          </Box>
-        )}
-      </Box>
-    </Box>
-  );
-}
-
-/**
- * Chip showing visited status for multi-location experiences
- * (Individual locations are toggled via checkboxes in the expanded list)
- */
-function VisitedStatusButton({
-  visitedStatus,
-  visitedCount,
-  totalCount,
-}: {
-  visitedStatus: VisitedStatus;
-  visitedCount: number;
-  totalCount: number;
-}) {
-  const statusConfig = {
-    not_visited: { label: `0/${totalCount} Visited`, color: 'default' as const },
-    partial: { label: `${visitedCount}/${totalCount} Visited`, color: 'warning' as const },
-    visited: { label: 'All Visited', color: 'success' as const },
-  };
-
-  const config = statusConfig[visitedStatus];
-
-  return (
-    <Chip
-      size="small"
-      label={config.label}
-      color={config.color}
-      variant={visitedStatus === 'visited' ? 'filled' : 'outlined'}
-      icon={
-        visitedStatus === 'visited' ? <CheckCircleIcon /> :
-        visitedStatus === 'partial' ? <PartialIcon /> : undefined
-      }
-    />
   );
 }
